Adjustments for 8th Homework, reworked index.ts file, removed part related to cars and adjusted ReadMe
